opt: performance improvements and modularization
Cursor optimization:
- Smart cursor blinking: only blink when idle (1.5s threshold), keep solid during activity
- Reduces unnecessary visual flicker during active use

Memory and rendering optimizations:
- Replace texture cache HashMap with LruCache (capacity 100) to prevent unbounded memory growth
- Adaptive frame budget (8KB-128KB) that dynamically adjusts based on frame time
- Scroll handling: merged 4 terminal locks into 1 for better lock contention

Zero-allocation improvements:
- key_to_string() returns &'static str (no allocations, full key coverage)
- build_keybinding_string() uses SmallVec<[&str; 5]> for modifier tracking
- keyboard_input_buffer reused across frames

Font and config optimizations:
- Lazy-initialize font detection with once_cell (removes blocking fc-list at startup)
- AVAILABLE_FONTS and MONOSPACE_FONTS are now static globals

Safety improvements:
- Add SAFETY documentation comments to all unsafe blocks in pty.rs, shell.rs, session_persistence.rs
- Clarifies justification for each unsafe operation

Compile-time optimizations:
- Add LTO, strip binary, set codegen-units=1 in release profile
- Add opt-level=1 for debug builds and dependencies

Modularization:
- Split main.rs (3743 → 3395 lines) into src/app/ submodules
- src/app/state.rs: TerminalApp struct definition
- src/app/events.rs: event handling and keybinding utilities
- src/app/rendering.rs: frame budget adaptation
- src/app/window.rs: config/session persistence methods
- src/app/tabs.rs, src/app/input.rs: module structure

Developer tools:
- Add criterion benchmarking framework with 4 initial benchmarks
- Expand debug_panel display: budget, pending output, texture cache metrics
- Add comprehensive README and REFACTORING documentation

+++ src/color.rs
@@ -53,7 +53,7 @@ pub fn color_256(idx: u8) -> Color32 {
             let r = (idx / 36) * 51;
             let g = ((idx % 36) / 6) * 51;
             let b = (idx % 6) * 51;
-            Color32::from_rgb(r as u8, g as u8, b as u8)
+            Color32::from_rgb(r, g, b)
         }
         232..=255 => {
             let gray = 8 + (idx - 232) * 10;
